refactor: MyPartyExerciseOrderType Enum을 Exercise 도메인 아래로 변경
