+++ src/main/java/umc/cockple/demo/domain/exercise/enums/MyPartyExerciseOrderType.java
@@ -1,4 +1,4 @@
-package umc.cockple.demo.global.enums;
+package umc.cockple.demo.domain.exercise.enums;
 
 import lombok.Getter;
